+++ src/components/feature/CartSection/index.tsx
@@ -1,20 +1,13 @@
 import * as S from "./CartSection.styles";
-import Card from "./CartProducts/Card";
 import Header from "./Header";
-import CheckBox from "../../common/CheckBox";
 import PriceSection from "./PriceSection";
 import useGetCartItem from "../../../hooks/useGetCartItem";
 import { useSelectedCart } from "../../../hooks/useSelectedCart";
+import CartList from "./CartList";
 
 const CartSection = () => {
   const { cartItems, refetch } = useGetCartItem();
-  const { selectedCartId, handleAllSelected, handleToggle, handleDelete } =
-    useSelectedCart(cartItems);
-
-  const isChecked = (id: number) => {
-    return selectedCartId?.some((item) => item === id);
-  };
-  const isAllChecked = selectedCartId?.length === cartItems?.length;
+  const { selectedCartId, setSelectedCartId } = useSelectedCart(cartItems);
 
   return (
     <S.Container>
@@ -29,24 +22,13 @@ const CartSection = () => {
             <S.Description>
               현재 {cartItems?.length}종류의 상품이 담겨있습니다.
             </S.Description>
-            <CheckBox
-              label="전체 선택"
-              isChecked={isAllChecked}
-              onChange={handleAllSelected}
-              testId="all-selected"
+
+            <CartList
+              cartItems={cartItems}
+              selectedCartId={selectedCartId}
+              setSelectedCartId={setSelectedCartId}
+              refetch={refetch}
             />
-            <S.CartList data-testid="cart-list">
-              {cartItems?.map((cartItem) => (
-                <Card
-                  key={cartItem.id}
-                  cartItem={cartItem}
-                  onRefetch={refetch}
-                  isChecked={isChecked(cartItem.id)}
-                  onToggle={() => handleToggle(cartItem.id)}
-                  onDeleteSelected={() => handleDelete(cartItem.id)}
-                />
-              ))}
-            </S.CartList>
 
             <PriceSection
               cartItems={cartItems}
