+++ src/hooks/useSelectedCart.ts
@@ -5,30 +5,6 @@ export const useSelectedCart = (cartItems: CartProduct[] | undefined) => {
   const [selectedCartId, setSelectedCartId] = useState<number[]>([]);
   const isSetting = useRef(false);
 
-  const handleAllSelected = () => {
-    const isAllChecked = selectedCartId?.length === cartItems?.length;
-    if (isAllChecked) {
-      setSelectedCartId([]);
-      return;
-    }
-    setSelectedCartId(cartItems?.map((item) => item.id) || []);
-  };
-
-  const handleToggle = (id: number) => {
-    // 선택이 안되어 있음 -> 선택됨
-    if (!selectedCartId?.find((item) => item === id)) {
-      setSelectedCartId((prev) => [...prev, id]);
-      return;
-    }
-    // 선택이 되어 있음 -> 선택안됨
-    setSelectedCartId(selectedCartId?.filter((cartId) => cartId !== id));
-  };
-
-  const handleDelete = (id: number) => {
-    if (!selectedCartId?.find((item) => item === id)) return;
-    setSelectedCartId(selectedCartId?.filter((cartId) => cartId !== id));
-  };
-
   useEffect(() => {
     if (cartItems && !isSetting.current) {
       const cartIdList = cartItems?.map((item) => item.id);
@@ -39,8 +15,6 @@ export const useSelectedCart = (cartItems: CartProduct[] | undefined) => {
 
   return {
     selectedCartId,
-    handleAllSelected,
-    handleToggle,
-    handleDelete,
+    setSelectedCartId,
   };
 };
